requirements.txt
djangorestframework==3.13.1
djoser==2.1.0
Pillow==9.1.1
